refactor(story_brief): extract dataset linting into linting module
