+++ tests/story_brief/test_coverage_gaps.py
@@ -8,7 +8,7 @@ import pytest
 
 from telegraphy.story_brief import generate_story_brief as story_brief
 from telegraphy.story_brief import data_io
-from telegraphy.story_brief.generate_story_brief import DatasetLintReport
+from telegraphy.story_brief.linting import DatasetLintReport
 from telegraphy.story_brief.partner_models import parse_partner_distribution_payload
